% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax_norm.R
\name{minmax_norm}
\alias{minmax_norm}
\title{MinMax normalization}
\usage{
minmax_norm(x)
}
\arguments{
\item{x, }{must be dbl or int.}
}
\description{
This function allows you to norm your data via Min and Max method.
}
\examples{
minmax_norm(iris$Sepal.Length)
}
\keyword{minmax}
\keyword{normalization}
