% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax_norm.R
\name{softmax_norm}
\alias{softmax_norm}
\title{softmax_norm normalization}
\usage{
softmax_norm(x)
}
\arguments{
\item{x, }{must be dbl or int.}
}
\description{
This function allows you to norm your data via SoftMax method.
}
\examples{
softmax_norm(iris$Sepal.Length)
}
\keyword{normalization}
\keyword{softmax}
