% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_norm.R
\name{zscore_norm}
\alias{zscore_norm}
\title{Z-score normalization}
\usage{
zscore_norm(x)
}
\arguments{
\item{x, }{must be dbl or int.}
}
\description{
This function allows you to norm your data via Z-score method.
}
\examples{
zscore_norm(iris$Sepal.Length)
}
\keyword{normalization}
\keyword{z}
\keyword{z-score}
